+++ src/main/java/persistence/entity/manager/EntityManagerImpl.java
@@ -9,7 +9,7 @@ import persistence.entity.persister.EntityPersister;
 
 import java.util.Optional;
 
-import static persistence.sql.dml.value.PrimaryKeyValue.getPrimaryKeyValue;
+import static persistence.sql.dml.clause.PrimaryKeyValue.getPrimaryKeyValue;
 
 public class EntityManagerImpl implements EntityManager {
     private final PersistenceContext persistenceContext;
