+++ src/main/java/persistence/sql/ddl/clause/PrimaryKeyClause.java
@@ -1,8 +1,8 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.clause;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
-import persistence.sql.dml.value.PrimaryKeyValue;
+import persistence.sql.dml.clause.PrimaryKeyValue;
 
 import java.lang.reflect.Field;
 import java.util.Map;
