+++ src/main/java/persistence/sql/ddl/querybuilder/DropQueryBuilder.java
@@ -1,4 +1,6 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.querybuilder;
+
+import persistence.sql.ddl.clause.TableClause;
 
 public class DropQueryBuilder {
     public static final String DROP_TABLE = "DROP TABLE IF EXISTS %s";
