+++ src/test/java/persistence/entity/entitypersister/EntityPersisterTest.java
@@ -10,7 +10,7 @@ import org.slf4j.LoggerFactory;
 import persistence.entity.loader.EntityLoader;
 import persistence.entity.persister.EntityPersister;
 import persistence.entity.testfixture.notcolumn.Person;
-import persistence.sql.ddl.CreateQueryBuilder;
+import persistence.sql.ddl.querybuilder.CreateQueryBuilder;
 
 import static persistence.sql.ddl.common.TestSqlConstant.DROP_TABLE_USERS;
