+++ src/test/java/persistence/entity/persistencecontext/PersistenceContextImplTest.java
@@ -10,7 +10,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import persistence.entity.testfixture.basic.Dog;
 import persistence.entity.testfixture.notcolumn.Person;
-import persistence.sql.ddl.CreateQueryBuilder;
+import persistence.sql.ddl.querybuilder.CreateQueryBuilder;
 
 import java.util.Optional;
