+++ src/test/java/persistence/repository/CustomJpaRepositoryTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import persistence.entity.testfixture.notcolumn.Person;
-import persistence.sql.ddl.CreateQueryBuilder;
+import persistence.sql.ddl.querybuilder.CreateQueryBuilder;
 
 import static persistence.sql.ddl.common.TestSqlConstant.DROP_TABLE_USERS;
