+++ src/test/java/persistence/sql/ddl/TableClauseTest.java
@@ -4,6 +4,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.entity.testfixture.annotated.Person;
+import persistence.sql.ddl.clause.TableClause;
+import persistence.sql.ddl.querybuilder.CreateQueryBuilder;
 
 class TableClauseTest {
     @Test
