+++ src/test/java/persistence/sql/dml/InsertQueryBuilderTest.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.entity.testfixture.notcolumn.Person;
+import persistence.sql.dml.querybuilder.InsertQueryBuilder;
 
 import java.util.List;
